requirements.txt
google-genai
requests
beautifulsoup4
google-auth
google-auth-oauthlib
google-auth-httplib2
